backend/dev-requirements.txt
# Development Dependencies

# Code Formatting
black>=24.0.0
isort>=5.13.0

# Linting
flake8>=7.0.0
flake8-django>=1.4
flake8-bugbear>=24.0.0
flake8-comprehensions>=3.15.0
pylint>=3.0.0
pylint-django>=2.5.5
ruff>=0.8.0

# Type Checking
mypy>=1.8.0
django-stubs>=5.2.9
djangorestframework-stubs>=3.16.8
types-requests>=2.32.0

# Testing
pytest>=8.0.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
factory-boy>=3.3.0
faker>=22.0.0

# Debugging
ipython>=8.20.0
ipdb>=0.13.13

# Code Quality
pre-commit>=3.6.0
bandit>=1.7.6

# Security
safety>=3.0.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0

# Editor Support
pylance>=2.0.0